+++ src/test/java/edu/kis/powp/jobs2d/events/SelectTestFigureCommandOptionListener.java
@@ -1,6 +1,6 @@
 package edu.kis.powp.jobs2d.events;
 
-import edu.kis.powp.command.*;
+import edu.kis.powp.jobs2d.command.*;
 import edu.kis.powp.jobs2d.drivers.DriverManager;
 
 import java.awt.event.*;
